+++ packages/loom-plugin-build-library-extended/src/index.ts
@@ -1,0 +1,4 @@
+export {
+  buildLibraryExtended,
+  buildLibraryExtendedWorkspace,
+} from './plugin-build-library-extended';