+++ packages/loom-plugin-build-library/src/plugin-rollup-config.ts
@@ -1,36 +1,157 @@
-import {rollupPlugins} from '@shopify/loom-plugin-package-build';
-import postcssShopify from '@shopify/postcss-plugin';
-import pluginGraphql from '@rollup/plugin-graphql';
-import json from '@rollup/plugin-json';
-import image from '@rollup/plugin-image';
-import svgr from '@svgr/rollup';
+import {
+  createProjectBuildPlugin,
+  Package,
+  Runtime,
+  DiagnosticError,
+} from '@shopify/loom';
+import babel from '@rollup/plugin-babel';
+import commonjs from '@rollup/plugin-commonjs';
+import nodeResolve from '@rollup/plugin-node-resolve';
+import externals from 'rollup-plugin-node-externals';
 
-import {styles} from './rollup/rollup-plugin-styles';
+import type {} from '@shopify/loom-plugin-babel';
 
-interface RollupConfigOptions {
-  readonly graphql?: boolean;
+declare module '@shopify/loom' {
+  interface BuildPackageTargetOptions {
+    rollupEsnext?: boolean;
+  }
 }
 
-export function rollupConfig({graphql = false}: RollupConfigOptions) {
-  return rollupPlugins((target) => {
-    const stylesConfig = target.options.rollupEsnext
-      ? {
-          mode: 'esnext',
-          plugins: [postcssShopify],
-        }
-      : {
-          mode: 'standalone',
-          output: 'styles.css',
-          plugins: [postcssShopify],
-        };
+interface RollupConfigOptions {
+  browserTargets: string;
+  nodeTargets: string;
+  commonjs: boolean;
+  esmodules: boolean;
+  esnext: boolean;
+}
 
-    return [
-      json(),
-      graphql && pluginGraphql(),
-      // TODO fix the svgr plugin casting as any when Plugin signature is compatible with rollup Plugin signature
-      svgr({include: '**/icons/*.svg', exclude: '', babel: true}) as any,
-      image({exclude: '**/icons/*.svg'}),
-      styles(stylesConfig),
-    ];
-  });
+export function rollupConfig(options: RollupConfigOptions) {
+  return createProjectBuildPlugin<Package>(
+    'Loom.BuildLibrary.RollupConfig',
+    ({hooks, project}) => {
+      // Define additional build variant to build esnext output
+      hooks.targets.hook((targets) => {
+        return targets.map((target) => {
+          return options.esnext && target.default
+            ? target.add({rollupEsnext: true})
+            : target;
+        });
+      });
+
+      // Define config for build variants
+      hooks.target.hook(async ({target, hooks}) => {
+        const isDefaultBuild = Object.keys(target.options).length === 0;
+        const isEsnextBuild = Boolean(target.options.rollupEsnext);
+        if (!(isDefaultBuild || isEsnextBuild)) {
+          return;
+        }
+
+        hooks.configure.hook(async (configuration) => {
+          configuration.rollupInput?.hook((input) => {
+            const inputEntries = [
+              ...target.project.entries,
+              ...target.project.binaries,
+            ].map(({root}) => require.resolve(root, {paths: [project.root]}));
+
+            if (inputEntries.length === 0) {
+              throw new DiagnosticError({
+                title: `No inputs found for "${project.name}".`,
+                suggestion: `Set a pkg.entry() in your loom.config. Use 'pkg.entry({root: './src/index'})" to use the index file`,
+              });
+            }
+
+            return input.concat(inputEntries);
+          });
+
+          configuration.rollupPlugins?.hook(async (plugins) => {
+            const babelConfig = await configuration.babelConfig?.run({});
+
+            const babelTargets: string[] = [];
+
+            if (isDefaultBuild) {
+              if (target.runtime.includes(Runtime.Browser)) {
+                babelTargets.push(options.browserTargets);
+              }
+              if (target.runtime.includes(Runtime.Node)) {
+                babelTargets.push(options.nodeTargets);
+              }
+            } else if (isEsnextBuild) {
+              babelTargets.push('last 1 chrome versions');
+            }
+
+            if (babelTargets.length === 0) {
+              throw new DiagnosticError({
+                title: `No targets found for "${project.name}".`,
+                suggestion: `Set a pkg.runtime() in your loom.config. Use "pkg.runtime(Runtime.Node)" for a node-only package. Use "pkg.runtime(Runtime.Node, Runtime.Browser)" for an isomorphic package that can be ran in node and the browser`,
+              });
+            }
+
+            const packagePath = (await project.fs.hasFile('./package.json'))
+              ? project.fs.resolvePath('./package.json')
+              : [];
+
+            return plugins.concat([
+              externals({
+                deps: true,
+                packagePath,
+              }),
+              nodeResolve({
+                extensions: ['.js', '.jsx', '.ts', '.tsx'],
+              }),
+              commonjs(),
+              babel({
+                ...babelConfig,
+                // Options specific to @rollup/plugin-babel, these can not be
+                // present on the `babelConfig` object
+                extensions: ['.js', '.jsx', '.ts', '.tsx'],
+                exclude: 'node_modules/**',
+                babelHelpers: 'bundled',
+                // Options that may be present on the `babelConfig` object but
+                // we want to override
+                envName: 'production',
+                // @ts-expect-error targets is a valid babel option but @types/babel__core doesn't know that yet
+                targets: babelTargets,
+              }),
+            ]);
+          });
+
+          configuration.rollupOutputs?.hook((outputs) => {
+            const additionalOutputs: typeof outputs = [];
+
+            if (isDefaultBuild) {
+              if (options.commonjs) {
+                additionalOutputs.push({
+                  format: 'cjs',
+                  dir: project.fs.buildPath('cjs'),
+                  preserveModules: true,
+                  entryFileNames: '[name][assetExtname].js',
+                  exports: 'named',
+                });
+              }
+
+              if (options.esmodules) {
+                additionalOutputs.push({
+                  format: 'esm',
+                  dir: project.fs.buildPath('esm'),
+                  preserveModules: true,
+                  entryFileNames: '[name][assetExtname].mjs',
+                });
+              }
+            } else if (isEsnextBuild) {
+              if (options.esmodules) {
+                additionalOutputs.push({
+                  format: 'esm',
+                  dir: project.fs.buildPath('esnext'),
+                  preserveModules: true,
+                  entryFileNames: '[name][assetExtname].esnext',
+                });
+              }
+            }
+
+            return outputs.concat(additionalOutputs);
+          });
+        });
+      });
+    },
+  );
 }